Rename Financial section to Rewards in My Dashboard
