+++ app/(saas)/my-dashboard/_translations/my-dashboard.translate.ts
@@ -1,5 +1,5 @@
-import enBudgetAvailable from "@/app/(saas)/my-dashboard/financial/_features/budget-available-cards/budget-available.en.json";
-import enFinancialColumnChart from "@/app/(saas)/my-dashboard/financial/_features/financial-column-chart/financial-column-chart.en.json";
+import enBudgetAvailable from "@/app/(saas)/my-dashboard/rewards/_features/budget-available-cards/budget-available.en.json";
+import enFinancialColumnChart from "@/app/(saas)/my-dashboard/rewards/_features/financial-column-chart/financial-column-chart.en.json";
 
 import enMyDashboard from "../_translations/my-dashboard.en.json";
